Code refactor with useGoogleConnectFlow.
