+++ js/src/components/google-account-card/connect-google-account-card.js
@@ -7,34 +7,13 @@ import { createInterpolateElement } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { glaData } from '.~/constants';
 import AccountCard, { APPEARANCE } from '.~/components/account-card';
-import useGoogleAuthorization from '.~/hooks/useGoogleAuthorization';
-import useDispatchCoreNotices from '.~/hooks/useDispatchCoreNotices';
 import AppButton from '.~/components/app-button';
 import readMoreLink from './read-more-link';
+import useGoogleConnectFlow from './use-google-connect-flow';
 
 export default function ConnectGoogleAccountCard( { disabled } ) {
-	const pageName = glaData.mcSetupComplete ? 'reconnect' : 'setup-mc';
-	const { createNotice } = useDispatchCoreNotices();
-	const [ fetchGoogleConnect, { loading, data } ] = useGoogleAuthorization(
-		pageName
-	);
-
-	const handleConnectClick = async () => {
-		try {
-			const { url } = await fetchGoogleConnect();
-			window.location.href = url;
-		} catch ( error ) {
-			createNotice(
-				'error',
-				__(
-					'Unable to connect your Google account. Please try again later.',
-					'google-listings-and-ads'
-				)
-			);
-		}
-	};
+	const [ handleConnect, { loading, data } ] = useGoogleConnectFlow();
 
 	return (
 		<AccountCard
@@ -70,7 +49,7 @@ export default function ConnectGoogleAccountCard( { disabled } ) {
 					loading={ loading || data }
 					eventName="gla_google_account_connect_button_click"
 					text={ __( 'Connect', 'google-listings-and-ads' ) }
-					onClick={ handleConnectClick }
+					onClick={ handleConnect }
 				/>
 			}
 		/>
