+++ js/src/components/google-account-card/connected-google-account-card.js
@@ -6,35 +6,14 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { glaData } from '.~/constants';
 import AccountCard, { APPEARANCE } from '.~/components/account-card';
 import AppButton from '.~/components/app-button';
 import ConnectedIconLabel from '.~/components/connected-icon-label';
-import useGoogleAuthorization from '.~/hooks/useGoogleAuthorization';
-import useDispatchCoreNotices from '.~/hooks/useDispatchCoreNotices';
 import Section from '.~/wcdl/section';
+import useGoogleConnectFlow from './use-google-connect-flow';
 
 export default function ConnectedGoogleAccountCard( { googleAccount } ) {
-	const pageName = glaData.mcSetupComplete ? 'reconnect' : 'setup-mc';
-	const { createNotice } = useDispatchCoreNotices();
-	const [ fetchGoogleConnect, { loading, data } ] = useGoogleAuthorization(
-		pageName
-	);
-
-	const handleConnect = async () => {
-		try {
-			const { url } = await fetchGoogleConnect();
-			window.location.href = url;
-		} catch ( error ) {
-			createNotice(
-				'error',
-				__(
-					'Unable to connect your Google account. Please try again later.',
-					'google-listings-and-ads'
-				)
-			);
-		}
-	};
+	const [ handleConnect, { loading, data } ] = useGoogleConnectFlow();
 
 	return (
 		<AccountCard
